requirements.txt
asgiref==3.8.0
crispy-bootstrap5==2024.2
Django==5.0.3
django-crispy-forms==2.1
django-taggit==5.0.1
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
